requirements.txt
selenium
discord.py-self
dotenv